+++ service/src/test/java/greencity/service/ubs/AddressServiceTest.java
@@ -109,7 +109,8 @@ class AddressServiceTest {
         List<LocationDto> locationDtos;
         List<DistrictDto> districtDtos;
         locationDtos = Arrays.asList(LocationDto.builder().id("1").build(), LocationDto.builder().id("2").build());
-        districtDtos = Arrays.asList(DistrictDto.builder().nameEn("District 1").build(), DistrictDto.builder().nameEn("District 2").build());
+        districtDtos = Arrays.asList(DistrictDto.builder().nameEn("District 1").build(),
+            DistrictDto.builder().nameEn("District 2").build());
         when(locationApiService.getAllDistrictsInCityByNames(anyString(), anyString())).thenReturn(locationDtos);
         when(mapper.map(any(LocationDto.class), eq(DistrictDto.class))).thenAnswer(i -> new DistrictDto());
         List<DistrictDto> results = addressService.getAllDistricts("region", "city");
@@ -122,7 +123,7 @@ class AddressServiceTest {
     @Test
     void checkOrderNotFound() {
         assertThrows(NotFoundException.class,
-                () -> addressService.getAddressByOrderId(10000000L));
+            () -> addressService.getAddressByOrderId(10000000L));
     }
 
     @Test
@@ -150,19 +151,22 @@ class AddressServiceTest {
     @Test
     void updateOrderAddressTest() {
         OrderAddress expected = ModelUtils.getOrderAddress();
-        OrderAddressExportDetailsDtoUpdate orderAddressExportDetailsDtoUpdate
-                = ModelUtils.getOrderAddressExportDetailsDtoUpdate();
+        OrderAddressExportDetailsDtoUpdate orderAddressExportDetailsDtoUpdate =
+            ModelUtils.getOrderAddressExportDetailsDtoUpdate();
         CreateAddressRequestDto createAddressRequestDto = ModelUtils.getAddressRequestDto();
         Address address = ModelUtils.getAddress();
         Region region = ModelUtils.getRegion();
         District district = ModelUtils.getDistrict();
         City city = ModelUtils.getCity();
-        when(mapper.map(orderAddressExportDetailsDtoUpdate, CreateAddressRequestDto.class)).thenReturn(createAddressRequestDto);
+        when(mapper.map(orderAddressExportDetailsDtoUpdate, CreateAddressRequestDto.class))
+            .thenReturn(createAddressRequestDto);
         when(mapper.map(any(OrderAddressExportDetailsDtoUpdate.class), eq(Address.class))).thenReturn(address);
         when(mapper.map(eq(address), eq(OrderAddress.class))).thenReturn(expected);
         when(regionRepository.findRegionByNameEnOrNameUk(anyString(), anyString())).thenReturn(Optional.of(region));
-        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString())).thenReturn(Optional.of(city));
-        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString())).thenReturn(Optional.of(district));
+        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.of(city));
+        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.of(district));
         when(baseEntityMapper.convert(eq(createAddressRequestDto), eq(District.class))).thenReturn(district);
         when(baseEntityMapper.convert(eq(createAddressRequestDto), eq(City.class))).thenReturn(city);
         addressService.updateOrderAddress(orderAddressExportDetailsDtoUpdate);
@@ -175,12 +179,14 @@ class AddressServiceTest {
     void updateOrderAddressIfNoRegionFoundTest() {
         CreateAddressRequestDto createAddressRequestDto = ModelUtils.getAddressRequestDto();
         Address address = ModelUtils.getAddress();
-        OrderAddressExportDetailsDtoUpdate orderAddressExportDetailsDtoUpdate
-                = ModelUtils.getOrderAddressExportDetailsDtoUpdate();
-        when(mapper.map(orderAddressExportDetailsDtoUpdate, CreateAddressRequestDto.class)).thenReturn(createAddressRequestDto);
+        OrderAddressExportDetailsDtoUpdate orderAddressExportDetailsDtoUpdate =
+            ModelUtils.getOrderAddressExportDetailsDtoUpdate();
+        when(mapper.map(orderAddressExportDetailsDtoUpdate, CreateAddressRequestDto.class))
+            .thenReturn(createAddressRequestDto);
         when(mapper.map(any(OrderAddressExportDetailsDtoUpdate.class), eq(Address.class))).thenReturn(address);
         when(regionRepository.findRegionByNameEnOrNameUk(anyString(), anyString())).thenReturn(Optional.empty());
-        assertThrows(BadRequestException.class, () -> addressService.updateOrderAddress(orderAddressExportDetailsDtoUpdate));
+        assertThrows(BadRequestException.class,
+            () -> addressService.updateOrderAddress(orderAddressExportDetailsDtoUpdate));
     }
 
     @Test
@@ -197,7 +203,8 @@ class AddressServiceTest {
         when(regionRepository.findRegionByNameEnOrNameUk(anyString(), anyString())).thenReturn(Optional.of(region));
         when(addressRepository.findAllByUserId(eq(user.getId()))).thenReturn(addresses);
         when(mapper.map(any(Address.class), eq(CreateAddressRequestDto.class))).thenReturn(createAddressRequestDto);
-        when(mapper.map(eq(createAddressRequestDto), eq(CreateAddressRequestDto.class))).thenReturn(createAddressRequestDto1);
+        when(mapper.map(eq(createAddressRequestDto), eq(CreateAddressRequestDto.class)))
+            .thenReturn(createAddressRequestDto1);
         addressService.saveCurrentAddressForOrder(createAddressRequestDto, user.getUuid());
         verify(userRepository, times(2)).findByUuid(anyString());
         verify(addressRepository, times(2)).findAllNonDeletedAddressesByUserId(eq(user.getId()));
@@ -215,15 +222,17 @@ class AddressServiceTest {
         when(regionRepository.findRegionByNameEnOrNameUk(anyString(), anyString())).thenReturn(Optional.of(region));
         when(addressRepository.findAllByUserId(eq(user.getId()))).thenReturn(addresses);
         when(mapper.map(any(Address.class), eq(CreateAddressRequestDto.class))).thenReturn(createAddressRequestDto);
-        when(mapper.map(eq(createAddressRequestDto), eq(CreateAddressRequestDto.class))).thenReturn(createAddressRequestDto);
-        assertThrows(BadRequestException.class, () -> addressService.saveCurrentAddressForOrder(createAddressRequestDto, user.getUuid()));
+        when(mapper.map(eq(createAddressRequestDto), eq(CreateAddressRequestDto.class)))
+            .thenReturn(createAddressRequestDto);
+        assertThrows(BadRequestException.class,
+            () -> addressService.saveCurrentAddressForOrder(createAddressRequestDto, user.getUuid()));
     }
 
     @Test
     void testUpdateAddressThrowsNotFoundOrderAddressException() {
         Order order = getOrder();
         assertThrows(NotFoundException.class,
-                () -> addressService.updateAddress(TEST_ORDER_ADDRESS_DTO_UPDATE, order, "abc"));
+            () -> addressService.updateAddress(TEST_ORDER_ADDRESS_DTO_UPDATE, order, "abc"));
     }
 
     @Test
@@ -240,7 +249,7 @@ class AddressServiceTest {
         when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
 
         BadRequestException exception = assertThrows(BadRequestException.class,
-                () -> addressService.makeAddressActual(firstAddressId, uuid));
+            () -> addressService.makeAddressActual(firstAddressId, uuid));
 
         assertEquals(CANNOT_MAKE_ACTUAL_DELETED_ADDRESS, exception.getMessage());
 
@@ -265,7 +274,7 @@ class AddressServiceTest {
         when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
 
         AccessDeniedException exception = assertThrows(AccessDeniedException.class,
-                () -> addressService.makeAddressActual(firstAddressId, uuid));
+            () -> addressService.makeAddressActual(firstAddressId, uuid));
 
         assertEquals(CANNOT_ACCESS_PERSONAL_INFO, exception.getMessage());
 
@@ -277,20 +286,20 @@ class AddressServiceTest {
     @Test
     void updateCurrentAddressForOrderWithNoExistingAddressTest() throws Exception {
         Method setLocations = AddressServiceImpl.class.getDeclaredMethod("setLocations",
-                CreateAddressRequestDto.class, Address.class);
+            CreateAddressRequestDto.class, Address.class);
         setLocations.setAccessible(true);
 
         Address address = getAddress();
         CreateAddressRequestDto dto = getAddressRequestToSaveDto();
         City city = ModelUtils.getCity();
         when(regionRepository.findRegionByNameEnOrNameUk(anyString(), anyString()))
-                .thenReturn(Optional.of(getRegion()));
+            .thenReturn(Optional.of(getRegion()));
         when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
-                .thenReturn(Optional.empty());
+            .thenReturn(Optional.empty());
         when(addressMapper.convert(any(), eq(City.class))).thenReturn(getCity());
         when(cityRepository.save(any())).thenReturn(getCity());
         when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
-                .thenReturn(Optional.empty());
+            .thenReturn(Optional.empty());
         when(addressMapper.convert(any(), eq(District.class))).thenReturn(getDistrict());
         when(baseEntityMapper.convert(eq(dto), eq(City.class))).thenReturn(city);
         setLocations.invoke(addressService, dto, address);
@@ -305,7 +314,7 @@ class AddressServiceTest {
     @Test
     void testAreAddressesEqual() throws Exception {
         Method areAddressesEqualMethod = AddressServiceImpl.class.getDeclaredMethod("areAddressesEqual",
-                CreateAddressRequestDto.class, CreateAddressRequestDto.class);
+            CreateAddressRequestDto.class, CreateAddressRequestDto.class);
         areAddressesEqualMethod.setAccessible(true);
 
         CreateAddressRequestDto address1 = getAddressRequestDtoReflection();
@@ -348,8 +357,8 @@ class AddressServiceTest {
         OrderAddressDtoRequest dtoRequest = getTestOrderAddressDtoRequest2();
         User user = getUser();
         Address address = getAddress(1L)
-                .setUser(user)
-                .setAddressStatus(AddressStatus.DELETED);
+            .setUser(user)
+            .setAddressStatus(AddressStatus.DELETED);
         CreateAddressRequestDto dto = getAddressRequestDto2();
 
         when(userRepository.findByUuid(anyString())).thenReturn(user);
@@ -371,8 +380,8 @@ class AddressServiceTest {
         OrderAddressDtoRequest dtoRequest = getTestOrderAddressDtoRequest2();
         User user = getUser();
         Address address = getAddress(1L)
-                .setUser(user)
-                .setAddressStatus(AddressStatus.DELETED);
+            .setUser(user)
+            .setAddressStatus(AddressStatus.DELETED);
         CreateAddressRequestDto dto = getAddressRequestDto();
 
         when(userRepository.findByUuid(anyString())).thenReturn(user);
@@ -381,10 +390,10 @@ class AddressServiceTest {
         when(mapper.map(any(), eq(CreateAddressRequestDto.class))).thenReturn(dto);
         when(mapper.map(any(), eq(Address.class))).thenReturn(getAddress());
         when(regionRepository.findRegionByNameEnOrNameUk(anyString(), anyString()))
-                .thenReturn(Optional.empty());
+            .thenReturn(Optional.empty());
 
         assertThrows(BadRequestException.class,
-                () -> addressService.updateCurrentAddressForOrder(dtoRequest, USER_UUID));
+            () -> addressService.updateCurrentAddressForOrder(dtoRequest, USER_UUID));
 
         verify(userRepository).findByUuid(anyString());
         verify(addressRepository).findById(anyLong());
@@ -420,27 +429,27 @@ class AddressServiceTest {
 
     private List<AddressDto> getTestAddressesDto() {
         AddressDto addressDto1 = AddressDto.builder().actual(true).id(13L).city("Kyiv").district("Svyatoshyn")
-                .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Peremohy av.")
-                .coordinates(new Coordinates(12.5, 34.5)).build();
+            .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Peremohy av.")
+            .coordinates(new Coordinates(12.5, 34.5)).build();
 
         AddressDto addressDto2 = AddressDto.builder().actual(true).id(42L).city("Lviv").district("Syhiv")
-                .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Lvivska st.")
-                .coordinates(new Coordinates(13.5, 36.5)).build();
+            .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Lvivska st.")
+            .coordinates(new Coordinates(13.5, 36.5)).build();
         return Arrays.asList(addressDto1, addressDto2);
     }
 
     private List<Address> getTestAddresses(User user) {
         Address address1 = Address.builder()
-                .addressStatus(AddressStatus.NEW).id(13L).city("Kyiv").district("Svyatoshyn")
-                .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Peremohy av.")
-                .user(user).actual(true).coordinates(new Coordinates(12.5, 34.5))
-                .build();
+            .addressStatus(AddressStatus.NEW).id(13L).city("Kyiv").district("Svyatoshyn")
+            .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Peremohy av.")
+            .user(user).actual(true).coordinates(new Coordinates(12.5, 34.5))
+            .build();
 
         Address address2 = Address.builder()
-                .addressStatus(AddressStatus.NEW).id(42L).city("Lviv").district("Syhiv")
-                .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Lvivska st.")
-                .user(user).actual(true).coordinates(new Coordinates(13.5, 36.5))
-                .build();
+            .addressStatus(AddressStatus.NEW).id(42L).city("Lviv").district("Syhiv")
+            .entranceNumber("1").houseCorpus("1").houseNumber("55").street("Lvivska st.")
+            .user(user).actual(true).coordinates(new Coordinates(13.5, 36.5))
+            .build();
 
         return Arrays.asList(address1, address2);
     }
@@ -462,9 +471,9 @@ class AddressServiceTest {
         when(mapper.map(any(), eq(Address.class))).thenReturn(addressToSave);
         when(regionRepository.findRegionByNameEnOrNameUk(any(), any())).thenReturn(Optional.of(getRegion()));
         when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
-                .thenReturn(Optional.of(getCity()));
+            .thenReturn(Optional.of(getCity()));
         when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
-                .thenReturn(Optional.of(getDistrict()));
+            .thenReturn(Optional.of(getDistrict()));
 
         var addressDto = addressDto();
         addressDto.setDistrict("Район");
@@ -472,13 +481,13 @@ class AddressServiceTest {
         when(mapper.map(addresses.getFirst(), AddressDto.class)).thenReturn(addressDto);
 
         OrderWithAddressesResponseDto actualWithSearchAddress =
-                addressService.saveCurrentAddressForOrder(createAddressRequestToSaveDto, uuid);
+            addressService.saveCurrentAddressForOrder(createAddressRequestToSaveDto, uuid);
 
         assertEquals(getAddressDtoResponse(), actualWithSearchAddress);
         assertEquals(createAddressRequestToSaveDto.getDistrict(),
-                actualWithSearchAddress.getAddressList().getFirst().getDistrict());
+            actualWithSearchAddress.getAddressList().getFirst().getDistrict());
         assertEquals(createAddressRequestToSaveDto.getDistrictEn(),
-                actualWithSearchAddress.getAddressList().getFirst().getDistrictEn());
+            actualWithSearchAddress.getAddressList().getFirst().getDistrictEn());
 
         verify(addressRepository).save(addressToSave);
 
@@ -529,7 +538,7 @@ class AddressServiceTest {
         when(addressRepository.findAllNonDeletedAddressesByUserId(user.getId())).thenReturn(addresses);
 
         BadRequestException exception = assertThrows(BadRequestException.class,
-                () -> addressService.saveCurrentAddressForOrder(createAddressRequestDto, uuid));
+            () -> addressService.saveCurrentAddressForOrder(createAddressRequestDto, uuid));
 
         assertEquals(ErrorMessage.NUMBER_OF_ADDRESSES_EXCEEDED, exception.getMessage());
     }
@@ -549,14 +558,14 @@ class AddressServiceTest {
         when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
         when(addressRepository.findAllNonDeletedAddressesByUserId(user.getId())).thenReturn(addresses);
         when(mapper.map(any(),
-                eq(CreateAddressRequestDto.class)))
-                .thenReturn(dtoRequest);
+            eq(CreateAddressRequestDto.class)))
+            .thenReturn(dtoRequest);
         when(addressRepository.findAllByUserId(user.getId())).thenReturn(addresses);
 
         dtoRequest.setPlaceId(null);
 
         BadRequestException exception = assertThrows(BadRequestException.class,
-                () -> addressService.saveCurrentAddressForOrder(createAddressRequestDto, uuid));
+            () -> addressService.saveCurrentAddressForOrder(createAddressRequestDto, uuid));
 
         assertEquals(ADDRESS_ALREADY_EXISTS, exception.getMessage());
 
@@ -581,35 +590,35 @@ class AddressServiceTest {
         when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
         when(addressRepository.findAllNonDeletedAddressesByUserId(user.getId())).thenReturn(addresses);
         when(addressRepository.findById(updateAddressRequestDto.getId()))
-                .thenReturn(Optional.ofNullable(addresses.getFirst()));
+            .thenReturn(Optional.ofNullable(addresses.getFirst()));
         when(mapper.map(any(),
-                eq(Address.class))).thenReturn(addresses.getFirst());
+            eq(Address.class))).thenReturn(addresses.getFirst());
         when(addressRepository.save(addresses.getFirst())).thenReturn(addresses.getFirst());
         when(regionRepository.findRegionByNameEnOrNameUk(any(), any())).thenReturn(Optional.of(getRegion()));
         when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
-                .thenReturn(Optional.of(getCity()));
+            .thenReturn(Optional.of(getCity()));
         when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
-                .thenReturn(Optional.of(getDistrict()));
+            .thenReturn(Optional.of(getDistrict()));
 
         var addressDto = addressDto();
         addressDto.setDistrict("Район");
         addressDto.setDistrictEn("District");
 
         when(mapper.map(addresses.getFirst(),
-                AddressDto.class))
-                .thenReturn(addressDto);
+            AddressDto.class))
+            .thenReturn(addressDto);
 
         OrderWithAddressesResponseDto actualWithSearchAddress =
-                addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid);
+            addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid);
 
         assertEquals(getAddressDtoResponse(), actualWithSearchAddress);
         assertEquals(updateAddressRequestDto.getDistrict(),
-                actualWithSearchAddress.getAddressList().getFirst().getDistrict());
+            actualWithSearchAddress.getAddressList().getFirst().getDistrict());
         assertEquals(updateAddressRequestDto.getDistrictEn(),
-                actualWithSearchAddress.getAddressList().getFirst().getDistrictEn());
+            actualWithSearchAddress.getAddressList().getFirst().getDistrictEn());
 
         OrderWithAddressesResponseDto actualWithoutSearchAddress =
-                addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid);
+            addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid);
 
         assertEquals(getAddressDtoResponse(), actualWithoutSearchAddress);
 
@@ -628,7 +637,7 @@ class AddressServiceTest {
         when(addressRepository.findById(firstAddressId)).thenReturn(Optional.empty());
 
         NotFoundException exception = assertThrows(NotFoundException.class,
-                () -> addressService.makeAddressActual(firstAddressId, uuid));
+            () -> addressService.makeAddressActual(firstAddressId, uuid));
 
         assertEquals(NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER + firstAddressId, exception.getMessage());
 
@@ -650,7 +659,7 @@ class AddressServiceTest {
         when(addressRepository.findByUserIdAndActualTrue(user.getId())).thenReturn(Optional.empty());
 
         NotFoundException exception = assertThrows(NotFoundException.class,
-                () -> addressService.makeAddressActual(firstAddressId, uuid));
+            () -> addressService.makeAddressActual(firstAddressId, uuid));
 
         assertEquals(ACTUAL_ADDRESS_NOT_FOUND, exception.getMessage());
 
@@ -673,24 +682,24 @@ class AddressServiceTest {
 
         when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
         when(addressRepository.findById(updateAddressRequestDto.getId()))
-                .thenReturn(Optional.ofNullable(addresses.getFirst()));
+            .thenReturn(Optional.ofNullable(addresses.getFirst()));
         when(addressRepository.findById(updateAddressRequestDto.getId()))
-                .thenReturn(Optional.ofNullable(addresses.getFirst()));
+            .thenReturn(Optional.ofNullable(addresses.getFirst()));
         when(regionRepository.findRegionByNameEnOrNameUk(any(), any())).thenReturn(Optional.of(getRegion()));
         when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
-                .thenReturn(Optional.of(getCity()));
+            .thenReturn(Optional.of(getCity()));
         when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
-                .thenReturn(Optional.of(getDistrict()));
+            .thenReturn(Optional.of(getDistrict()));
         when(mapper.map(any(),
-                eq(Address.class))).thenReturn(addresses.getFirst());
+            eq(Address.class))).thenReturn(addresses.getFirst());
 
         when(addressRepository.save(addresses.getFirst())).thenReturn(addresses.getFirst());
 
         OrderWithAddressesResponseDto actualWithSearchAddress =
-                addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid);
+            addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid);
 
         assertEquals(OrderWithAddressesResponseDto.builder().addressList(Collections.emptyList()).build(),
-                actualWithSearchAddress);
+            actualWithSearchAddress);
 
         verify(addressRepository, times(1)).save(addresses.getFirst());
         verify(addressRepository, times(1)).findById(anyLong());
@@ -713,14 +722,14 @@ class AddressServiceTest {
 
         when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
         when(addressRepository.findById(updateAddressRequestDto.getId()))
-                .thenReturn(Optional.ofNullable(addresses.getFirst()));
+            .thenReturn(Optional.ofNullable(addresses.getFirst()));
         when(mapper.map(any(),
-                eq(CreateAddressRequestDto.class)))
-                .thenReturn(dtoRequest);
+            eq(CreateAddressRequestDto.class)))
+            .thenReturn(dtoRequest);
         when(addressRepository.findAllByUserId(user.getId())).thenReturn(addresses);
 
         BadRequestException exception = assertThrows(BadRequestException.class,
-                () -> addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid));
+            () -> addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid));
 
         assertEquals(ADDRESS_ALREADY_EXISTS, exception.getMessage());
     }
@@ -746,7 +755,7 @@ class AddressServiceTest {
         when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
 
         AccessDeniedException exception = assertThrows(AccessDeniedException.class,
-                () -> addressService.updateCurrentAddressForOrder(dtoRequest, uuid));
+            () -> addressService.updateCurrentAddressForOrder(dtoRequest, uuid));
 
         assertEquals(CANNOT_ACCESS_PERSONAL_INFO, exception.getMessage());
     }
@@ -768,7 +777,7 @@ class AddressServiceTest {
 
         when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
         when(addressRepository.findAnyByUserIdAndAddressStatusNotDeleted(user.getId()))
-                .thenReturn(Optional.of(secondAddress));
+            .thenReturn(Optional.of(secondAddress));
         doReturn(new OrderWithAddressesResponseDto()).when(addressServiceSpy).findAllAddressesForCurrentOrder(uuid);
         when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
         addressServiceSpy.deleteCurrentAddressForOrder(firstAddressId, uuid);
@@ -823,7 +832,7 @@ class AddressServiceTest {
         when(addressRepository.findById(updateAddressRequestDto.getId())).thenReturn(Optional.empty());
 
         NotFoundException exception = assertThrows(NotFoundException.class,
-                () -> addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid));
+            () -> addressService.updateCurrentAddressForOrder(updateAddressRequestDto, uuid));
 
         assertEquals(NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER + updateAddressRequestDto.getId(), exception.getMessage());
     }
@@ -837,7 +846,7 @@ class AddressServiceTest {
         when(addressRepository.findById(addressId)).thenReturn(Optional.empty());
 
         NotFoundException exception = assertThrows(NotFoundException.class,
-                () -> addressServiceSpy.deleteCurrentAddressForOrder(addressId, "qwe"));
+            () -> addressServiceSpy.deleteCurrentAddressForOrder(addressId, "qwe"));
 
         assertEquals(NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER + addressId, exception.getMessage());
 
@@ -860,7 +869,7 @@ class AddressServiceTest {
         when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
 
         AccessDeniedException exception = assertThrows(AccessDeniedException.class,
-                () -> addressServiceSpy.deleteCurrentAddressForOrder(firstAddressId, uuid));
+            () -> addressServiceSpy.deleteCurrentAddressForOrder(firstAddressId, uuid));
 
         assertEquals(CANNOT_DELETE_ADDRESS, exception.getMessage());
 
@@ -884,7 +893,7 @@ class AddressServiceTest {
         when(addressRepository.findById(firstAddressId)).thenReturn(Optional.of(firstAddress));
 
         BadRequestException exception = assertThrows(BadRequestException.class,
-                () -> addressServiceSpy.deleteCurrentAddressForOrder(firstAddressId, uuid));
+            () -> addressServiceSpy.deleteCurrentAddressForOrder(firstAddressId, uuid));
 
         assertEquals(CANNOT_DELETE_ALREADY_DELETED_ADDRESS, exception.getMessage());
 
@@ -929,19 +938,19 @@ class AddressServiceTest {
 
     @Test
     void mapUpdatedOrderAddressFieldsTest() throws Exception {
-        Method mapUpdatedOrderAddressFieldsMethod = AddressServiceImpl.class.getDeclaredMethod("mapUpdatedOrderAddressFields",
+        Method mapUpdatedOrderAddressFieldsMethod =
+            AddressServiceImpl.class.getDeclaredMethod("mapUpdatedOrderAddressFields",
                 OrderAddress.class, OrderAddress.class, String.class);
         mapUpdatedOrderAddressFieldsMethod.setAccessible(true);
         Location location = ModelUtils.getLocation();
         String comment = "New comment test";
-        OrderAddress actual = OrderAddress.builder().
-                location(location)
-                .id(1L)
-                .actual(false)
-                .addressComment("No comments")
-                .coordinates(location.getCoordinates())
-                .addressStatus(AddressStatus.DELETED)
-                .build();
+        OrderAddress actual = OrderAddress.builder().location(location)
+            .id(1L)
+            .actual(false)
+            .addressComment("No comments")
+            .coordinates(location.getCoordinates())
+            .addressStatus(AddressStatus.DELETED)
+            .build();
         OrderAddress expected = ModelUtils.getOrderAddress();
         mapUpdatedOrderAddressFieldsMethod.invoke(addressService, expected, actual, comment);
         assertEquals(expected.getLocation(), actual.getLocation());
@@ -951,6 +960,7 @@ class AddressServiceTest {
         assertEquals(expected.getCoordinates(), actual.getCoordinates());
         assertEquals(expected.getAddressStatus(), actual.getAddressStatus());
     }
+
     @Test
     void addressUpdateIfPresentTest() {
         Order order = ModelUtils.getOrder();
@@ -965,16 +975,118 @@ class AddressServiceTest {
         when(orderAddressRepository.findById(anyLong())).thenReturn(Optional.of(orderAddress));
         when(orderRepository.findById(anyLong())).thenReturn(Optional.of(order));
         when(regionRepository.findRegionByNameEnOrNameUk(anyString(), anyString())).thenReturn(Optional.of(region));
-        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString())).thenReturn(Optional.empty());
-        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString())).thenReturn(Optional.empty());
+        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.empty());
+        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.empty());
         when(mapper.map(any(OrderAddressExportDetailsDtoUpdate.class), eq(Address.class))).thenReturn(address);
-        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString())).thenReturn(Optional.of(city));
-        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString())).thenReturn(Optional.of(district));
+        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.of(city));
+        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.of(district));
         when(mapper.map(any(Address.class), eq(OrderAddress.class))).thenReturn(orderAddress);
-        when(mapper.map(any(OrderAddress.class), eq(OrderAddressDtoResponse.class))).thenReturn(orderAddressDtoResponse);
+        when(mapper.map(any(OrderAddress.class), eq(OrderAddressDtoResponse.class)))
+            .thenReturn(orderAddressDtoResponse);
         addressService.addressUpdate(addressDto, "email@gmail.com");
         verify(orderRepository, times(1)).findById(order.getId());
         verify(orderAddressRepository, times(1)).save(any());
 
+    }
+
+    @Test
+    void saveCurrentAddressForOrderForAddressesBelongToKyivEnTest() {
+        User user = getUserForCreate();
+        List<Address> addresses = user.getAddresses();
+        addresses.getFirst().setActual(false);
+        addresses.getFirst().setAddressStatus(AddressStatus.NEW);
+
+        String uuid = user.getUuid();
+        CreateAddressRequestDto createAddressRequestToSaveDto = getAddressWithKyivRegionToSaveRequestDto();
+        Address addressToSave = getAddress();
+
+        when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
+        when(addressRepository.findAllNonDeletedAddressesByUserId(user.getId())).thenReturn(addresses);
+        when(regionRepository.findRegionByNameEnOrNameUk(any(), any())).thenReturn(Optional.of(getRegion()));
+        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.of(getCity()));
+        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.of(getDistrict()));
+
+        when(mapper.map(any(), eq(Address.class))).thenReturn(addressToSave);
+        when(mapper.map(addresses.getFirst(), AddressDto.class)).thenReturn(addressWithKyivRegionDto());
+
+        OrderWithAddressesResponseDto actualWithSearchAddress =
+            addressService.saveCurrentAddressForOrder(createAddressRequestToSaveDto, uuid);
+
+        assertEquals(KYIV_REGION_EN, actualWithSearchAddress.getAddressList().getFirst().getRegionEn());
+
+        verify(userRepository, times(2)).findByUuid(user.getUuid());
+        verify(addressRepository, times(2)).findAllNonDeletedAddressesByUserId(user.getId());
+
+        verify(mapper, times(1)).map(any(), eq(Address.class));
+        verify(mapper).map(any(), eq(Address.class));
+        verify(mapper).map(addresses.getFirst(), AddressDto.class);
+        verify(regionRepository).findRegionByNameEnOrNameUk(anyString(), anyString());
+        verify(cityRepository).findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString());
+        verify(districtRepository).findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString());
+        verify(addressRepository).save(addressToSave);
+    }
+
+    @Test
+    void saveCurrentAddressForOrderForAddressesBelongToKyivUaTest() {
+        User user = getUserForCreate();
+        List<Address> addresses = user.getAddresses();
+        addresses.getFirst().setActual(false);
+        addresses.getFirst().setAddressStatus(AddressStatus.NEW);
+
+        String uuid = user.getUuid();
+        CreateAddressRequestDto createAddressRequestToSaveDto = getAddressWithKyivRegionToSaveRequestDto();
+        Address addressToSave = getAddress();
+
+        when(userRepository.findByUuid(user.getUuid())).thenReturn(user);
+        when(addressRepository.findAllNonDeletedAddressesByUserId(user.getId())).thenReturn(addresses);
+        when(regionRepository.findRegionByNameEnOrNameUk(any(), any())).thenReturn(Optional.of(getRegion()));
+        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.of(getCity()));
+        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.of(getDistrict()));
+        when(mapper.map(any(), eq(Address.class))).thenReturn(addressToSave);
+        when(mapper.map(addresses.getFirst(), AddressDto.class)).thenReturn(addressWithKyivRegionDto());
+
+        OrderWithAddressesResponseDto actualWithSearchAddress =
+            addressService.saveCurrentAddressForOrder(createAddressRequestToSaveDto, uuid);
+
+        assertEquals(KYIV_REGION_UA, actualWithSearchAddress.getAddressList().getFirst().getRegion());
+
+        verify(userRepository, times(2)).findByUuid(user.getUuid());
+        verify(addressRepository, times(2)).findAllNonDeletedAddressesByUserId(user.getId());
+
+        verify(mapper, times(1)).map(any(), eq(Address.class));
+        verify(mapper).map(any(), eq(Address.class));
+        verify(mapper).map(addresses.getFirst(), AddressDto.class);
+        verify(regionRepository).findRegionByNameEnOrNameUk(anyString(), anyString());
+        verify(cityRepository).findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString());
+        verify(districtRepository).findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString());
+        verify(addressRepository).save(addressToSave);
+    }
+
+    @Test
+    void testUpdateOrderAddress() {
+        Address addressToSave = getAddress();
+
+        when(mapper.map(TEST_ORDER_ADDRESS_DTO_UPDATE, CreateAddressRequestDto.class))
+            .thenReturn(TEST_CREATE_ADDRESS_DTO);
+        when(mapper.map(any(), eq(Address.class))).thenReturn(addressToSave);
+        when(regionRepository.findRegionByNameEnOrNameUk(any(), any())).thenReturn(Optional.of(getRegion()));
+        when(cityRepository.findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.of(getCity()));
+        when(districtRepository.findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString()))
+            .thenReturn(Optional.of(getDistrict()));
+
+        addressService.updateOrderAddress(TEST_ORDER_ADDRESS_DTO_UPDATE);
+
+        verify(regionRepository).findRegionByNameEnOrNameUk(anyString(), anyString());
+        verify(cityRepository).findCityByRegionIdAndNameUkAndNameEn(anyLong(), anyString(), anyString());
+        verify(districtRepository).findDistrictByCityIdAndNameEnOrNameUk(anyLong(), anyString(), anyString());
     }
 }
